Preparing for add new `InternalAffairs/DynamicPredicateName` cop
Refactoring with extract module.
