+++ lib/rubocop.rb
@@ -94,6 +94,7 @@ require_relative 'rubocop/cop/mixin/configurable_enforced_style'
 require_relative 'rubocop/cop/mixin/configurable_formatting'
 require_relative 'rubocop/cop/mixin/configurable_naming'
 require_relative 'rubocop/cop/mixin/configurable_numbering'
+require_relative 'rubocop/cop/mixin/configurable_predicate_naming'
 require_relative 'rubocop/cop/mixin/def_node'
 require_relative 'rubocop/cop/mixin/documentation_comment'
 require_relative 'rubocop/cop/mixin/duplication'
