+++ lib/rubocop/cop/naming/predicate_name.rb
@@ -18,12 +18,14 @@ module RuboCop
       #   # good
       #   def value? ...
       class PredicateName < Cop
+        include ConfigurablePredicateNaming
+
         def on_def(node)
           predicate_prefixes.each do |prefix|
             method_name = node.method_name.to_s
-            next unless method_name.start_with?(prefix)
-            next if method_name == expected_name(method_name, prefix)
-            next if predicate_whitelist.include?(method_name)
+
+            next if valid_method_name?(method_name, prefix)
+
             add_offense(
               node,
               location: :name,
@@ -32,34 +34,6 @@ module RuboCop
           end
         end
         alias on_defs on_def
-
-        private
-
-        def expected_name(method_name, prefix)
-          new_name = if prefix_blacklist.include?(prefix)
-                       method_name.sub(prefix, '')
-                     else
-                       method_name.dup
-                     end
-          new_name << '?' unless method_name.end_with?('?')
-          new_name
-        end
-
-        def message(method_name, new_name)
-          "Rename `#{method_name}` to `#{new_name}`."
-        end
-
-        def prefix_blacklist
-          cop_config['NamePrefixBlacklist']
-        end
-
-        def predicate_prefixes
-          cop_config['NamePrefix']
-        end
-
-        def predicate_whitelist
-          cop_config['NameWhitelist']
-        end
       end
     end
   end
